backend/requirements.txt
# aigua.tv WebUI Backend Requirements

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data Validation
pydantic>=2.5.0

# WebSocket Support
websockets>=12.0

# Redis Support (optional - will fallback to in-memory if not available)
redis>=5.0.1

# File Upload Support
python-multipart>=0.0.6

# Configuration
pyyaml>=6.0.0

# Existing dependencies (from main project)
# These should already be installed if running aigua.tv CLI
# openai>=1.0.0
# tmdbv3api>=2.0.0
